requirements.txt
pyyaml>=3.10
paho-mqtt
pyA20
click
python-daemon